test(refresh): classify terminal outcome tests
